requirements.txt
opencv-python == 4.6.0.66
opencv-contrib-python == 4.6.0.66
pandas == 1.5.0
scikit-learn == 1.1.2
matplotlib == 3.6.0
scikit-image == 0.19.3
seaborn == 0.12.0
tqdm == 4.64.1
SimpleITK-SimpleElastix == 2.0.0rc2.dev909
PyYAML == 6.0
unzip == 1.0.0
antspyx == 0.3.4
